【测试成功】成功显示职位信息但无法筛选，其他栏目功能正常

+++ src/main/java/com/platform/ats/entity/job/dto/JobInfoQueryDto.java
@@ -19,8 +19,8 @@ public class JobInfoQueryDto implements Serializable {
     private Long companyId;
     
     // 薪资筛选
-    private Integer salaryMin;
-    private Integer salaryMax;
+    private BigDecimal salaryMin;
+    private BigDecimal salaryMax;
     
     // 排序字段
     private String orderBy;
